requirements.txt
-i https://pypi.org/simple
asgiref==3.2.10; python_version >= '3.5'
dj-database-url==0.5.0
dj-email-url==1.0.1
django-cache-url==3.1.2
django-filter==2.4.0
django==3.1.2
djangorestframework==3.12.1
environs[django]==8.0.0
gunicorn==20.0.4
markdown==3.3.3
marshmallow==3.8.0; python_version >= '3.5'
psycopg2==2.8.6
python-dotenv==0.14.0
pytz==2020.1
sqlparse==0.4.1; python_version >= '3.5'
whitenoise==5.2.0
